V0.9 Scalable dimensions/area
The contour of the area went from 4,000 pixels to a scalable area depending on how big the dimensions of the given image are.
This area can be changed my changing the variable: percentage_area. It's currently at 0.1, which means the area has to exceed at least 10% of the entire image.